chore: moves scm to plugin folder
